refactor(core): split database.rs (1419L) into focused sub-modules (Phase 6)
Split monolithic database.rs into database/ module:
- mod.rs: Database struct, open/close, config
- collection_ops.rs: collection CRUD
- persistence.rs: save/load/flush
- query_engine.rs: VelesQL execution
- stats.rs: statistics and diagnostics
- training.rs: quantization training
